backend/requirements.txt
fastapi==0.115.0
uvicorn==0.34.0
pydantic==2.10.3
pydantic-settings==2.7.0
sqlalchemy==2.0.36
asyncpg==0.31.2
alembic==1.14.0
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
better-exception==0.3.3
qdrant-client==1.12.1
openai==1.7.2
python-dotenv==1.0.1
httpx==0.28.1